Add files via upload
This is a map for a contest that I participated in for a fan-game called Megaman 8-bit Deathmatch under a pseudonym or a nickname (Artyom), It was developed over a month after the deadline and it was made for different game modes like Death matches, Teams, Capture the flag, etc. It got a 4th place over 10 contestants.